Adicionando Iluminacao e Camera

+++ rugosidadePG2/rugosidadePG2/Point.swift
@@ -9,7 +9,7 @@
 import Foundation
 
 struct Point {
-    var x: Float
-    var y: Float
-    var z: Float?
+    var x: Double
+    var y: Double
+    var z: Double?
 }
